Modularized the Test Step Accordions
